# src/SFA.DAS.Location.Api/swagger_v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LocationApi",
    "version": "v1"
  },
  "paths": {
    "/api/Addresses": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minMatch",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetAddressesListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAddressesListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAddressesListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Addresses/ByCoordinates": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestedPlace"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestedPlace"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestedPlace"
                }
              }
            }
          }
        }
      }
    },
    "/api/Locations": {
      "get": {
        "tags": [
          "Locations"
        ],
        "parameters": [
          {
            "name": "locationName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorityName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/Postcodes": {
      "get": {
        "tags": [
          "Postcodes"
        ],
        "parameters": [
          {
            "name": "postcode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/Postcodes/outcode": {
      "get": {
        "tags": [
          "Postcodes"
        ],
        "parameters": [
          {
            "name": "outcode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/Postcodes/bulk": {
      "post": {
        "tags": [
          "Postcodes"
        ],
        "parameters": [
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Search": {
      "get": {
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "results",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLocationsListResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Geometry": {
        "type": "object",
        "properties": {
          "coordinates": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetAddressesListItem": {
        "type": "object",
        "properties": {
          "uprn": {
            "type": "string",
            "nullable": true
          },
          "organisation": {
            "type": "string",
            "nullable": true
          },
          "premises": {
            "type": "string",
            "nullable": true
          },
          "thoroughfare": {
            "type": "string",
            "nullable": true
          },
          "locality": {
            "type": "string",
            "nullable": true
          },
          "postTown": {
            "type": "string",
            "nullable": true
          },
          "county": {
            "type": "string",
            "nullable": true
          },
          "postcode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "addressLine3": {
            "type": "string",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "match": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetAddressesListResponse": {
        "type": "object",
        "properties": {
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetAddressesListItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetLocationsListItem": {
        "type": "object",
        "properties": {
          "localAuthorityName": {
            "type": "string",
            "nullable": true
          },
          "countyName": {
            "type": "string",
            "nullable": true
          },
          "locationName": {
            "type": "string",
            "nullable": true
          },
          "postcode": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/Geometry"
          },
          "districtName": {
            "type": "string",
            "nullable": true
          },
          "outcode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetLocationsListResponse": {
        "type": "object",
        "properties": {
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetLocationsListItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SuggestedPlace": {
        "type": "object",
        "properties": {
          "uprn": {
            "type": "string",
            "nullable": true
          },
          "buildingName": {
            "type": "string",
            "nullable": true
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "addressLine3": {
            "type": "string",
            "nullable": true
          },
          "postcode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}